fix: login check message move to hooks
