+++ pages/interestVideo/index.tsx
@@ -6,9 +6,7 @@ import { useQuery } from "react-query";
 import NoResult from "components/custom/no-result";
 import { Loading } from "components/custom";
 import { QUERY_KEY } from "lib/query-key";
-import Router from "next/router";
-import { useDispatch } from "react-redux";
-import modalSlice from "reducers/modal";
+import { useCheck } from "hooks/use-check";
 
 const TITLE = "관심 영상 없음";
 const CONTENTS = `관심 영상이 없어요 :( 
@@ -16,7 +14,9 @@ const CONTENTS = `관심 영상이 없어요 :(
 
 const InterestVideo: NextPageWithLayout = () => {
   const [noResult, setNoResult] = useState(true);
-  const dispatch = useDispatch();
+
+  const { checkModal } = useCheck();
+
   const { isLoading, data } = useQuery(
     QUERY_KEY.likesAll.key,
     QUERY_KEY.likesAll.api,
@@ -31,20 +31,7 @@ const InterestVideo: NextPageWithLayout = () => {
         });
       },
       onError: () => {
-        const email = sessionStorage.getItem("email");
-        if (!email) {
-          Router.replace("/");
-          dispatch(modalSlice.actions.isError({ isError: true }));
-          dispatch(modalSlice.actions.open({}));
-          dispatch(
-            modalSlice.actions.setErrorMessage({
-              errorMessage: "로그인이 필요합니다.",
-            })
-          );
-          setTimeout(() => {
-            dispatch(modalSlice.actions.close({}));
-          }, 3000);
-        }
+        checkModal();
       },
     }
   );
